+++ edc_auth/__init__.py
@@ -1,6 +1,5 @@
 from .codenames import data_manager, screening
-from .get_default_codenames_by_group import get_default_codenames_by_group
-from .group_names import (
+from .default_group_names import (
     ACCOUNT_MANAGER,
     ADMINISTRATION,
     AE,
@@ -24,6 +23,5 @@ from .group_names import (
     SCREENING,
     SITE_DATA_MANAGER,
     TMG,
-    UNBLINDING_REQUESTORS,
-    UNBLINDING_REVIEWERS,
 )
+from .default_groups import default_groups
