+++ edc_auth/apps.py
@@ -5,22 +5,17 @@ from django.core.checks.registry import register
 from django.core.management.color import color_style
 from django.db.models.signals import post_migrate
 
-from .role_names import groups_by_role_name, role_names
+from .auth_updater import AuthUpdater
 from .system_checks import edc_check
-from .update_roles import update_roles
 
 style = color_style()
 
 
-def post_migrate_user_roles(sender=None, **kwargs):  # noqa
-    """Update Role model with EDC defaults.
+def post_migrate_user_groups_and_roles(sender=None, **kwargs):  # noqa
+    """Update Groups, Role model with EDC defaults."""
+    from django.apps import apps as django_apps
 
-    To add custom roles, register this in your main
-    app with additional role_names and groups_by_role_name.
-    """
-    update_roles(
-        groups_by_role_name=groups_by_role_name, role_names=role_names, verbose=True
-    )
+    AuthUpdater(apps=django_apps)
 
 
 class AppConfig(DjangoAppConfig):
@@ -28,7 +23,10 @@ class AppConfig(DjangoAppConfig):
     verbose_name = "Edc Authentication"
 
     def ready(self):
-        post_migrate.connect(post_migrate_user_roles, sender=self)
-        register(edc_check)
+        from .site_auths import site_auths
+
         sys.stdout.write(f"Loading {self.verbose_name} ...\n")
+        site_auths.autodiscover()
+        post_migrate.connect(post_migrate_user_groups_and_roles, sender=self)
+        register(edc_check)
         sys.stdout.write(f" Done loading {self.verbose_name}.\n")
