+++ edc_auth/auth_updater/__init__.py
@@ -1,0 +1,1 @@
+from .auth_updater import AuthUpdater
