+++ edc_auth/codename_tuples/dashboard_tuples.py
@@ -17,11 +17,11 @@ dashboard_tuples = [
     ("edc_dashboard.view_ae_listboard", "Can view AE listboard"),
     ("edc_dashboard.view_enrolment_listboard", "Can view Enrolment listboard"),
     ("edc_dashboard.view_export_dashboard", "Can view Export Dashboard"),
-    ("edc_dashboard.view_my_ae_listboard", "Can view my AE listboard"),
-    ("edc_dashboard.view_my_enrolment_listboard", "Can view my Enrolment listboard"),
-    ("edc_dashboard.view_my_screening_listboard", "Can view my Screening listboard"),
-    ("edc_dashboard.view_my_subject_listboard", "Can view my Subject listboard"),
-    ("edc_dashboard.view_my_tmg_listboard", "Can view my TMG Listboard"),
+    # ("edc_dashboard.view_my_ae_listboard", "Can view my AE listboard"),
+    # ("edc_dashboard.view_my_enrolment_listboard", "Can view my Enrolment listboard"),
+    # ("edc_dashboard.view_my_screening_listboard", "Can view my Screening listboard"),
+    # ("edc_dashboard.view_my_subject_listboard", "Can view my Subject listboard"),
+    # ("edc_dashboard.view_my_tmg_listboard", "Can view my TMG Listboard"),
     ("edc_dashboard.view_screening_listboard", "Can view Screening listboard"),
     ("edc_dashboard.view_subject_listboard", "Can view Subject listboard"),
     ("edc_dashboard.view_tmg_listboard", "Can view TMG Listboard"),
