+++ edc_auth/constants.py
@@ -1,17 +1,1 @@
-ACCOUNT_MANAGER_ROLE = "account_manager"
-AUDITOR_ROLE = "auditor"
-CLINICIAN_ROLE = "clinician"
-CLINICIAN_SUPER_ROLE = "clinician_super"
-CUSTOM_ROLE = "custom"
-DATA_EXPORTER_ROLE = "data_exporter"
-DATA_MANAGER_ROLE = "data_manager"
-LAB_TECHNICIAN_ROLE = "laboratory_technician"
-NURSE_ROLE = "research_nurse"
-PHARMACIST_ROLE = "pharmacist"
-RANDO_VIEW_ROLE = "rando_view_role"
-SITE_COORDINATOR = "site_coordinator"
-SITE_DATA_MANAGER_ROLE = "site_data_manager"
-SITE_PHARMACIST_ROLE = "site_pharmacist"
-STAFF_ROLE = "staff"
-STATISTICIAN_ROLE = "statistician"
-TMG_ROLE = "tmg"
+
