+++ edc_auth/default_group_names.py
@@ -24,5 +24,3 @@ SITE_DATA_MANAGER = "SITE_DATA_MANAGER"
 STATISTICIAN = "STATISTICIAN"
 TMG = "TMG"
 TMG_REVIEW = "TMG_REVIEW"
-UNBLINDING_REQUESTORS = "UNBLINDING_REQUESTORS"
-UNBLINDING_REVIEWERS = "UNBLINDING_REVIEWERS"
