+++ edc_auth/default_role_names.py
@@ -1,0 +1,37 @@
+ACCOUNT_MANAGER_ROLE = "account_manager"
+AUDITOR_ROLE = "auditor"
+CLINICIAN_ROLE = "clinician"
+CLINICIAN_SUPER_ROLE = "clinician_super"
+CUSTOM_ROLE = "custom"
+DATA_EXPORTER_ROLE = "data_exporter"
+DATA_MANAGER_ROLE = "data_manager"
+LAB_TECHNICIAN_ROLE = "laboratory_technician"
+NURSE_ROLE = "research_nurse"
+PHARMACIST_ROLE = "pharmacist"
+RANDO_VIEW_ROLE = "rando_view_role"
+SITE_COORDINATOR = "site_coordinator"
+SITE_DATA_MANAGER_ROLE = "site_data_manager"
+SITE_PHARMACIST_ROLE = "site_pharmacist"
+STAFF_ROLE = "staff"
+STATISTICIAN_ROLE = "statistician"
+TMG_ROLE = "tmg"
+
+default_role_names = {
+    ACCOUNT_MANAGER_ROLE: "Account Manager",
+    AUDITOR_ROLE: "Auditor",
+    CLINICIAN_ROLE: "Clinician",
+    CLINICIAN_SUPER_ROLE: "Clinician Super",
+    CUSTOM_ROLE: "Custom ...",
+    DATA_EXPORTER_ROLE: "Data Exporter",
+    DATA_MANAGER_ROLE: "Data Manager",
+    LAB_TECHNICIAN_ROLE: "Laboratory Technician",
+    NURSE_ROLE: "Research Nurse",
+    PHARMACIST_ROLE: "Pharmacist",
+    RANDO_VIEW_ROLE: "Randomization view",
+    SITE_COORDINATOR: "Site Coordinator",
+    SITE_DATA_MANAGER_ROLE: "Site Data Manager",
+    SITE_PHARMACIST_ROLE: "Site Pharmacist",
+    STATISTICIAN_ROLE: "Statistician",
+    STAFF_ROLE: "Staff",
+    TMG_ROLE: "TMG (External Review)",
+}
