+++ edc_auth/forms.py
@@ -4,9 +4,9 @@ from django.contrib.auth.forms import UserChangeForm as BaseForm
 from django.utils.safestring import mark_safe
 from edc_randomization.blinding import is_blinded_user
 
-from .group_names import PHARMACY
+from .default_group_names import PHARMACY
+from .default_role_names import PHARMACIST_ROLE
 from .models import UserProfile
-from .role_names import PHARMACIST_ROLE
 
 
 class UserChangeForm(BaseForm):
